requirements.txt
Jinja2==3.0.1
PyGithub==1.55
PySimpleGUI==4.45.0
PyYAML==5.4.1
requests==2.26.0
schema==0.7.4
docx-mailmerge==0.5.0
